chore: repo consolidation — deduplicate detections, extract SIEM module, update coverage
- Delete 9 duplicate detection files from execution/ (T1070.001, T1219, T1566.004
  were incorrectly placed; canonical copies remain in defense_evasion/ and initial_access/)
- Extract SIEM deployment logic into shared orchestration/siem.py module
  (~170 lines removed from blue_team_agent.py, credentials now read from config.yml)
- Backfill detection-request files for T1547.001 and T1134.001 (pre-pipeline detections)
- Backfill test_results for T1059.001, T1547.001, T1134.001 (required by state schema)
- Update coverage/attack-matrix.md: 8 deployed + 1 monitoring + 1 in progress
- Update coverage/detection-backlog.md: mark 5 new deployments, update build-next list
- Update coverage/data-sources.md: reflect current SIEM status
- Update STATUS.md with accurate pipeline state (9 detections, 38% coverage)
- Add .gitignore entries for runtime artifacts (budget-log.jsonl, audit-log.jsonl)
- Remove obsolete TODO-lab-workflow.md (all items completed)
