+++ autonomous/detection-requests/t1547_001.yml
@@ -1,0 +1,29 @@
+technique_id: T1547.001
+title: Registry Run Key Persistence
+status: DEPLOYED
+priority: high
+mitre_tactic: persistence
+requested_by: manual
+requested_date: '2026-03-01T00:00:00Z'
+intel_report: threat-intel/fawkes/fawkes-ttp-mapping.md
+scenario_file: ''
+sigma_rule: detections/persistence/t1547_001_registry_run_key.yml
+compiled_lucene: ''
+compiled_spl: ''
+deployed_date: '2026-03-01T00:00:00Z'
+elastic_rule_id: bddcc2fe-4e9a-48af-a674-756c4bcb1ecd
+splunk_saved_search: Registry Run Key T1547.001
+last_quality_review: ''
+quality_score: 1.0
+fp_rate: 0.0
+tp_rate: 1.0
+alert_volume_24h: 0
+cost_estimate: low
+auto_deploy_eligible: true
+fawkes_commands:
+- persist
+changelog:
+- date: '2026-03-01T00:00:00Z'
+  agent: manual
+  action: created
+  details: Manually deployed before Patronus pipeline. TP 2/2, FP 0/6088.
