+++ autonomous/orchestration/agents/blue_team_agent.py
@@ -11,16 +11,15 @@ Called by agent_runner.py. Implements run(state_manager) interface.
 import datetime
 import json
 import subprocess
-import textwrap
-import urllib.request
-import urllib.error
 from pathlib import Path
 from uuid import uuid4
+
 
 import yaml
 
 from orchestration.state import StateManager
 from orchestration import learnings
+from orchestration.siem import deploy_to_siems, TACTIC_MAP, SEVERITY_MAP
 
 AUTONOMOUS_DIR = Path(__file__).resolve().parent.parent.parent
 REPO_ROOT = AUTONOMOUS_DIR.parent
@@ -42,20 +41,6 @@ def _today() -> str:
     return datetime.datetime.now(datetime.timezone.utc).strftime("%Y-%m-%d")
 
 
-# ─── MITRE Tactic Mapping ──────────────────────────────────────────
-TACTIC_MAP = {
-    "execution": ("Execution", "TA0002"),
-    "persistence": ("Persistence", "TA0003"),
-    "privilege_escalation": ("Privilege Escalation", "TA0004"),
-    "defense_evasion": ("Defense Evasion", "TA0005"),
-    "credential_access": ("Credential Access", "TA0006"),
-    "discovery": ("Discovery", "TA0007"),
-    "lateral_movement": ("Lateral Movement", "TA0008"),
-    "collection": ("Collection", "TA0009"),
-    "command_and_control": ("Command and Control", "TA0011"),
-    "initial_access": ("Initial Access", "TA0001"),
-}
-
 # Sigma logsource categories by Sysmon EID
 SYSMON_CATEGORIES = {
     "1": "process_creation",
@@ -68,14 +53,6 @@ SYSMON_CATEGORIES = {
     "22": "dns_query",
 }
 
-# Risk score mapping
-SEVERITY_MAP = {
-    "informational": (21, "low"),
-    "low": (47, "medium"),
-    "medium": (73, "high"),
-    "high": (73, "high"),
-    "critical": (99, "critical"),
-}
 
 
 def load_scenario(scenario_path: str) -> dict | None:
@@ -402,179 +379,6 @@ def assess_quality(metrics: dict) -> str:
         return "human_review"
     else:
         return "needs_rework"
-
-
-# ─── SIEM Deployment ──────────────────────────────────────────────
-
-ES_URL = "http://localhost:9200"
-KIBANA_URL = "http://localhost:5601"
-ES_AUTH = ("elastic", "changeme")
-SPLUNK_URL = "https://localhost:8089"
-SPLUNK_AUTH = ("admin", "BlueTeamLab1!")
-
-
-def _check_siem_available(siem: str) -> bool:
-    """Check if a SIEM is reachable."""
-    try:
-        if siem == "elastic":
-            req = urllib.request.Request(f"{ES_URL}/_cluster/health")
-            creds = f"{ES_AUTH[0]}:{ES_AUTH[1]}"
-            import base64
-            req.add_header("Authorization", "Basic " + base64.b64encode(creds.encode()).decode())
-            with urllib.request.urlopen(req, timeout=5) as resp:
-                return resp.status == 200
-        elif siem == "splunk":
-            import ssl
-            ctx = ssl.create_default_context()
-            ctx.check_hostname = False
-            ctx.verify_mode = ssl.CERT_NONE
-            req = urllib.request.Request(f"{SPLUNK_URL}/services/server/health")
-            creds = f"{SPLUNK_AUTH[0]}:{SPLUNK_AUTH[1]}"
-            import base64
-            req.add_header("Authorization", "Basic " + base64.b64encode(creds.encode()).decode())
-            with urllib.request.urlopen(req, context=ctx, timeout=5) as resp:
-                return resp.status == 200
-    except Exception:
-        return False
-    return False
-
-
-def deploy_to_elastic(request: dict, lucene_query: str, sigma_rule: dict) -> dict | None:
-    """Deploy detection rule to Elastic Security Detection Engine."""
-    if not _check_siem_available("elastic"):
-        print("    [blue-team] Elastic not available — skipping Elastic deployment")
-        return None
-
-    tid = request["technique_id"]
-    tactic_key = request.get("mitre_tactic", "execution")
-    tactic_name, tactic_id = TACTIC_MAP.get(tactic_key, ("Execution", "TA0002"))
-    title = sigma_rule.get("title", f"Detection for {tid}")
-    severity = sigma_rule.get("level", "high")
-    risk_score, _ = SEVERITY_MAP.get(severity, (73, "high"))
-
-    rule_id = str(uuid4())
-    rule_payload = {
-        "rule_id": rule_id,
-        "name": title,
-        "description": sigma_rule.get("description", f"Detects {tid}"),
-        "type": "query",
-        "query": lucene_query,
-        "index": ["sim-*"],
-        "language": "lucene",
-        "severity": severity,
-        "risk_score": risk_score,
-        "interval": "5m",
-        "from": "now-6m",
-        "enabled": True,
-        "tags": [tid, tactic_name],
-        "threat": [{
-            "framework": "MITRE ATT&CK",
-            "tactic": {"id": tactic_id, "name": tactic_name,
-                       "reference": f"https://attack.mitre.org/tactics/{tactic_id}/"},
-            "technique": [{"id": tid.split(".")[0], "name": request.get("title", tid),
-                           "reference": f"https://attack.mitre.org/techniques/{tid.replace('.', '/')}/"}],
-        }],
-    }
-
-    import base64
-    data = json.dumps(rule_payload).encode()
-    req = urllib.request.Request(
-        f"{KIBANA_URL}/api/detection_engine/rules",
-        data=data, method="POST",
-        headers={
-            "kbn-xsrf": "true",
-            "Content-Type": "application/json",
-            "Authorization": "Basic " + base64.b64encode(f"{ES_AUTH[0]}:{ES_AUTH[1]}".encode()).decode(),
-        },
-    )
-
-    try:
-        with urllib.request.urlopen(req, timeout=15) as resp:
-            result = json.loads(resp.read())
-            print(f"    [blue-team] Elastic rule created: {result.get('id', rule_id)}")
-            return {"rule_id": rule_id, "elastic_id": result.get("id", "")}
-    except urllib.error.HTTPError as e:
-        body = e.read().decode()[:300]
-        print(f"    [blue-team] Elastic deploy failed ({e.code}): {body}")
-        return None
-    except Exception as e:
-        print(f"    [blue-team] Elastic deploy error: {e}")
-        return None
-
-
-def deploy_to_splunk(request: dict, spl_query: str, sigma_rule: dict) -> dict | None:
-    """Deploy detection as a Splunk saved search with alerting."""
-    if not _check_siem_available("splunk"):
-        print("    [blue-team] Splunk not available — skipping Splunk deployment")
-        return None
-
-    tid = request["technique_id"]
-    title = sigma_rule.get("title", f"Detection for {tid}")
-    search_name = f"{request.get('title', tid)} {tid}"
-    severity = sigma_rule.get("level", "high")
-    sev_map = {"informational": "1", "low": "2", "medium": "3", "high": "4", "critical": "5"}
-    alert_severity = sev_map.get(severity, "4")
-
-    # Build SPL search — wrap the transpiled query with index and table
-    full_spl = f'index=sysmon {spl_query} | table _time host user process.name process.command_line process.executable'
-
-    import base64, ssl, urllib.parse
-    ctx = ssl.create_default_context()
-    ctx.check_hostname = False
-    ctx.verify_mode = ssl.CERT_NONE
-
-    params = urllib.parse.urlencode({
-        "name": search_name,
-        "search": full_spl,
-        "is_scheduled": "1",
-        "cron_schedule": "*/5 * * * *",
-        "alert_type": "number of events",
-        "alert_comparator": "greater than",
-        "alert_threshold": "0",
-        "alert.severity": alert_severity,
-        "output_mode": "json",
-    }).encode()
-
-    req = urllib.request.Request(
-        f"{SPLUNK_URL}/servicesNS/admin/search/saved/searches",
-        data=params, method="POST",
-        headers={
-            "Authorization": "Basic " + base64.b64encode(f"{SPLUNK_AUTH[0]}:{SPLUNK_AUTH[1]}".encode()).decode(),
-        },
-    )
-
-    try:
-        with urllib.request.urlopen(req, context=ctx, timeout=15) as resp:
-            result = json.loads(resp.read())
-            name = result.get("entry", [{}])[0].get("name", search_name)
-            print(f"    [blue-team] Splunk saved search created: {name}")
-            return {"search_name": name}
-    except urllib.error.HTTPError as e:
-        body = e.read().decode()[:300]
-        # If already exists, that's OK
-        if "already exists" in body.lower():
-            print(f"    [blue-team] Splunk saved search already exists: {search_name}")
-            return {"search_name": search_name}
-        print(f"    [blue-team] Splunk deploy failed ({e.code}): {body}")
-        return None
-    except Exception as e:
-        print(f"    [blue-team] Splunk deploy error: {e}")
-        return None
-
-
-def deploy_to_siems(request: dict, lucene: str, spl: str, sigma_rule: dict) -> dict:
-    """Deploy to all available SIEMs. Returns deployment results."""
-    results = {}
-
-    elastic_result = deploy_to_elastic(request, lucene, sigma_rule)
-    if elastic_result:
-        results["elastic"] = elastic_result
-
-    splunk_result = deploy_to_splunk(request, spl, sigma_rule)
-    if splunk_result:
-        results["splunk"] = splunk_result
-
-    return results
 
 
 def author_and_validate(request: dict, state_manager: StateManager, run_id: str) -> dict:
