--- detections/execution/compiled/t1070_001.spl
@@ -1,1 +1,0 @@
-event.code="1" process.name IN ("powershell.exe", "wevtutil.exe") process.command_line IN ("powershell.exe -Command \"Clear-EventLog -LogName Security\"", "wevtutil.exe cl Security") NOT process.name="wevtutil.exe"
