--- detections/execution/compiled/t1219.spl
@@ -1,1 +1,0 @@
-event.code="1" process.name IN ("AnyDesk.exe", "splashtop.exe") process.executable IN ("C:\\Users\\agarcia\\AppData\\Local\\Temp\\splashtop.exe", "C:\\Users\\agarcia\\Downloads\\AnyDesk.exe") process.parent.name="explorer.exe" NOT process.name="TeamViewer.exe"
