--- detections/execution/compiled/t1566_004.lucene
@@ -1,1 +1,0 @@
-event.code:4624 AND source.ip:203.0.113.42 AND winlog.logon.type:RemoteInteractive
