--- detections/execution/compiled/t1566_004.spl
@@ -1,1 +1,0 @@
-event.code="4624" source.ip="203.0.113.42" winlog.logon.type="RemoteInteractive"
