--- detections/execution/t1070_001.yml
@@ -1,32 +1,0 @@
-title: Clear Windows Event Logs [T1070.001]
-id: d391d142-725b-4883-a091-8ec260ee776c
-status: experimental
-description: Detects clear windows event logs. Detection should flag wevtutil cl and Clear-EventLog targeting Security/System
-  logs. Exclude wevtutil qe (query, not clear).
-author: blue-team-agent
-date: '2026-03-06'
-modified: '2026-03-06'
-references:
-- https://attack.mitre.org/techniques/T1070/001/
-tags:
-- attack.execution
-- attack.t1070.001
-logsource:
-  category: process_creation
-  product: windows
-detection:
-  selection:
-    event.code: '1'
-    process.name:
-    - powershell.exe
-    - wevtutil.exe
-    process.command_line:
-    - powershell.exe -Command "Clear-EventLog -LogName Security"
-    - wevtutil.exe cl Security
-  condition: selection and not filter_legitimate
-  filter_legitimate:
-    process.name:
-    - wevtutil.exe
-falsepositives:
-- Detection should flag wevtutil cl and Clear-EventLog targeting Security/System logs. Exclude wevtutil qe (query, not clear).
-level: high
