--- detections/execution/t1219.yml
@@ -1,34 +1,0 @@
-title: Remote Access Software [T1219]
-id: 3d42cb29-8849-4180-bb3f-d15b3f876b3c
-status: experimental
-description: Detects remote access software. Detection should flag RMM tools (AnyDesk, Splashtop, TeamViewer, ScreenConnect)
-  running from non-standard paths (Downloads, Temp, AppData). Exclude approved installations in Program Files.
-author: blue-team-agent
-date: '2026-03-06'
-modified: '2026-03-06'
-references:
-- https://attack.mitre.org/techniques/T1219/
-tags:
-- attack.execution
-- attack.t1219
-logsource:
-  category: process_creation
-  product: windows
-detection:
-  selection:
-    event.code: '1'
-    process.name:
-    - AnyDesk.exe
-    - splashtop.exe
-    process.executable:
-    - C:\Users\agarcia\AppData\Local\Temp\splashtop.exe
-    - C:\Users\agarcia\Downloads\AnyDesk.exe
-    process.parent.name: explorer.exe
-  condition: selection and not filter_legitimate
-  filter_legitimate:
-    process.name:
-    - TeamViewer.exe
-falsepositives:
-- Detection should flag RMM tools (AnyDesk, Splashtop, TeamViewer, ScreenConnect) running from non-standard paths (Downloads,
-  Temp, AppData). Exclude approved installations in Program Files.
-level: high
