--- detections/execution/t1566_004.yml
@@ -1,26 +1,0 @@
-title: Spearphishing Voice (Vishing) [T1566.004]
-id: 86b95ca9-c3a3-4b90-b1dd-1fd29db77677
-status: experimental
-description: "Detects spearphishing voice (vishing). Limited endpoint telemetry for vishing. Detection relies on correlation:\
-  \ MFA reset + immediate external login. Sysmon alone insufficient \u2014 needs identity provider logs."
-author: blue-team-agent
-date: '2026-03-06'
-modified: '2026-03-06'
-references:
-- https://attack.mitre.org/techniques/T1566/004/
-tags:
-- attack.execution
-- attack.t1566.004
-logsource:
-  category: authentication
-  product: windows
-detection:
-  selection:
-    event.code: '4624'
-    source.ip: 203.0.113.42
-    winlog.logon.type: RemoteInteractive
-  condition: selection
-falsepositives:
-- "Limited endpoint telemetry for vishing. Detection relies on correlation: MFA reset + immediate external login. Sysmon alone\
-  \ insufficient \u2014 needs identity provider logs."
-level: high
