+++ threat-intel/digest.md
@@ -13,3 +13,17 @@
 - [CISA Advisory: Scattered Spider Social Engineering and RMM Abuse](https://www.cisa.gov/news-events/cybersecurity-advisories/aa23-320a) — 4 techniques
 
 ---
+
+
+## Intel Run — 2026-03-06
+
+**Reports processed**: 1
+**New techniques found**: 0
+**Detection requests created**: 0
+**Techniques skipped (existing)**: 4
+**Fawkes overlap**: 0
+
+### Reports
+- [CISA Advisory: Scattered Spider Social Engineering and RMM Abuse](https://www.cisa.gov/news-events/cybersecurity-advisories/aa23-320a) — 4 techniques
+
+---
